Convert to a Maven project
Instead of requiring libs/ folder with jars, the dependencies will be
managed using a pom.xml through Maven. Additionally, the namespaces were
moved around as well as updating the namespaces for the dependencies
