+++ .gitignore
@@ -1,17 +1,20 @@
-#personal
-/.idea/
-/testing/
-/recordings/
-token
-shark_token
-settings.json
-*.wav
-*.mp3
+#
+# Standard Maven .gitignore
+#
+target/
+pom.xml.tag
+pom.xml.releaseBackup
+pom.xml.versionsBackup
+pom.xml.next
+release.properties
+dependency-reduced-pom.xml
+buildNumber.properties
+.mvn/timing.properties
 
+#
 # IntelliJ
-/out/
-/.idea/
-/src/META-INF/
+#
 *.iml
-*.ipr
-*.iws
+.idea/*
+!.idea/runConfigurations/
+
