+++ .travis.yml
@@ -1,0 +1,4 @@
+language: java
+jdk: oraclejdk8
+after_success:
+- mvn coveralls:report
