--- src/com/DiscordEcho/Commands/Command.java
@@ -1,11 +1,0 @@
-package com.DiscordEcho.Commands;
-
-import net.dv8tion.jda.core.events.message.guild.GuildMessageReceivedEvent;
-
-public interface Command {
-    Boolean called(String[] args, GuildMessageReceivedEvent e);
-    void action(String[] args, GuildMessageReceivedEvent e);
-    String usage(String prefix);
-    String descripition();
-    void executed(boolean success, GuildMessageReceivedEvent e);
-}
