+++ src/main/java/tech/gdragon/App.java
@@ -1,0 +1,17 @@
+package tech.gdragon;
+
+/**
+ * Hello world!
+ */
+public final class App {
+    private App() {
+    }
+
+    /**
+     * Says hello to the world.
+     * @param args The arguments of the program.
+     */
+    public static void main(String[] args) {
+        System.out.println("Hello World!");
+    }
+}
