+++ src/test/java/tech/gdragon/AppTest.java
@@ -1,0 +1,18 @@
+package tech.gdragon;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+/**
+ * Unit test for simple App.
+ */
+public class AppTest {
+    /**
+     * Rigorous Test.
+     */
+    @Test
+    public void testApp() {
+        assertTrue(true);
+    }
+}
